--- packages/@lwc/engine-server/src/__tests__/fixtures/wire/method/expected.html
@@ -1,11 +1,0 @@
-<x-wire>
-  <template shadowrootmode="open">
-    wired field value: {
-    key2: [
-        fixed,
-        array
-    ],
-    key1: foo
-}
-  </template>
-</x-wire>
