+++ packages/@lwc/ssr-compiler/src/compile-template/transformers/component/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2024, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: MIT
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+ */
+
+export * from './component';
+export * from './lwc-component';
